Don't update avail_cpus based on --cpus-per-task when --cpus-per-gpu
Use of --cpus-per-gpu is mutually exclusive with --cpus-per-task after
5ad29f8af96, which means that the value in cpus_per_task is not a
cpus_per_task_orig (which we don't have in gres_mc_data_t), but
cpus_per_task subjected to a sainty default setting.

On the other hand if cpus_per_gres was used (and is exclusive with
--cpus-per-task) we already have the logic that set number of
available_cpus based in gres_select_filter_sock_core.

Bug 15182