Attempt equal placing of cpus on GRES in first_pass
If we're in first_pass and the request is for -n --gpus, we can attempt
using same number of cpus per GRES for optimal task placement.

This makes the following algorithm aware of number of cpus potentially
required for task on socket to reduce number of GPUs used per socket if
we know that we won't be able to get enough CPUs on the socket to use
higher number of GPUs on the socket.

Bug 15403