No functional change - create infrastructure for next commit
This adds an assoc_locked argument to _adjust_limit_usage and its
callers that can be used to pass the information that assoc locks
required for the function are already owned by the thread.
The commit doesn't use the functionality anywhere - it's just to make
the future commits easier to understand.

Bug 15718