gres_sched_add/_set_job_bits1 - don't recalculate GRES available on node
The logic in gres_select_filter_sock_core checks how many GRES devices
we can use on the node in this pass. This is more complicated than what
we checked in other places. We shouldn't recalculate it, but just use
what gres_filter_sock_core decided in the following logic.

Bug 15403