Just copy new sysinfo instead of checking values
This changes the behavior in terms of last_node_update. Previously it
was under conditional check, but in reality chances that freemem and 3
digits of load5 where the same for next registration message is
incidental.

At the same time, if one is using sinfo -i2 he's rather not interested
in getting freeram/loads.