_srun_slurm_set_addr - fallback to logic before 11712ce7099
If we fail to find node in the configuration we should fallback to just
using a string we had stored in step_ptr. This may happen if
NodeHostname configured differs from gethostname() result on host, but
matches one of the aliases.

Another case is srun running outside of the cluster - submit host, which
usually is not configured as Slurm node.

Bug 16401